Update start.sh
Got rid of the pid files. 
A container restart (using --with-gmond) did not restart gmond. I suspect that the gmetad start up is killing the gmond that was just started because the 'old' pid of the gmetad is the same as the new pid of gmond. 

Removing the pid file stuff makes the container start correctly